+++ src/types/tspexpress.d.ts
@@ -1,0 +1,2 @@
+/// <reference path="../../node_modules/tspexpress/src/types/golbal.d.ts" />
+declare module 'tspexpress';