+++ src/utils/common.ts
@@ -3,5 +3,3 @@ const path = require('path');
 export const BASE_DIR = path.resolve(__dirname, '../../MyBlog/dist');
 
 export const DEFAULT_FILE = 'index.html';
-
-export const MODEL_DATA_DIR = path.resolve(__dirname, '../../debug');