+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.0.1 (2020-06-13)
+
+
+### Features
+
+* **add tspexpress:** add tspexpress and nodemon ([b05d5b5](https://github.com/hblvsjtu/MyBlogServer/commit/b05d5b52b84551b82d2422126d36fbe5fdbbddbc))
+* **defaultDir:** make default Dir to visit ([9bbd649](https://github.com/hblvsjtu/MyBlogServer/commit/9bbd649fc9f6f0088e9c74b1643066c83504c64c))
+* **fileRead:** fileRead method ([3a6df2f](https://github.com/hblvsjtu/MyBlogServer/commit/3a6df2f342b450d203528270eb703d4cb74c02e3))
+* **init:** initial myBlogServer project ([707487a](https://github.com/hblvsjtu/MyBlogServer/commit/707487acbb132ab1c523736a1c718ce40651ff7a))
+* **InterceptManager:** InterceptManager with promise API ([bbadc8f](https://github.com/hblvsjtu/MyBlogServer/commit/bbadc8fe54b77f3f1072a507a2915c2fc4e9ac19))
+* **nodemon:** build nodemon script ([282c9bc](https://github.com/hblvsjtu/MyBlogServer/commit/282c9bc0badcfb28c2e06e16b5e206b826d5094f))
+* **staticPath:** buildStaticPath method ([9475453](https://github.com/hblvsjtu/MyBlogServer/commit/9475453e85c3d72de5e91551178a43ad784cbb0e))
+
+
+
 ## 0.0.1 (2020-06-07)